Extract originator_link to new Store Credit helper
This commit extracts the necessary logic to generate originator links
for Store Credit-associated objects to a new helper, so that it can be
more easily re-used. This is to accomodate the new
StoreCreditLedgerEntry model, which also has a polymorphic originator
association. This allows the helper code to be more easily re-used
between StoreCreditEvents and the aforementioned ledger entries.